+++ packages/model/src/standard/elements/root-node.element.ts
@@ -14,7 +14,7 @@ import {
 } from "../../elements/index.js";
 
 export const RootNodeDt = DeviceType(
-    { name: "RootNode", id: 0x16, classification: "node" },
+    { name: "RootNode", id: 0x16, classification: "node", composition: "full-family" },
     Requirement(
         { name: "Descriptor", id: 0x1d, element: "serverCluster" },
         Requirement({ name: "DeviceTypeList", default: [ { deviceType: 22, revision: 4 } ], element: "attribute" })